requirements.txt
adal==1.2.1
aniso8601==4.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cryptography==2.5
flasgger==0.9.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
google-auth==1.6.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0
kubernetes==8.0.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==6.0.0
oauthlib==3.0.1
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.14.11
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.55.0
Werkzeug==0.14.1
docker-compose==1.24.1
